Remove files

+++ application/src/main/java/it/smartcommunitylabdhub/core/services/ArtifactServiceImpl.java
@@ -18,6 +18,7 @@ import it.smartcommunitylabdhub.commons.services.FilesInfoService;
 import it.smartcommunitylabdhub.commons.services.RelationshipsAwareEntityService;
 import it.smartcommunitylabdhub.commons.services.SpecRegistry;
 import it.smartcommunitylabdhub.core.components.infrastructure.specs.SpecValidator;
+import it.smartcommunitylabdhub.core.components.security.UserAuthenticationHelper;
 import it.smartcommunitylabdhub.core.models.builders.artifact.ArtifactEntityBuilder;
 import it.smartcommunitylabdhub.core.models.entities.AbstractEntity_;
 import it.smartcommunitylabdhub.core.models.entities.ArtifactEntity;
@@ -406,10 +407,30 @@ public class ArtifactServiceImpl
     }
 
     @Override
-    public void deleteArtifact(@NotNull String id) {
+    public void deleteArtifact(@NotNull String id, @Nullable Boolean cascade) {
         log.debug("delete artifact with id {}", String.valueOf(id));
+
         try {
-            entityService.delete(id);
+            Artifact artifact = entityService.find(id);
+            if (artifact != null) {
+                if (Boolean.TRUE.equals(cascade)) {
+                    //files
+                    log.debug("cascade delete files for artifact with id {}", String.valueOf(id));
+
+                    //extract path from spec
+                    ArtifactBaseSpec spec = new ArtifactBaseSpec();
+                    spec.configure(artifact.getSpec());
+
+                    String path = spec.getPath();
+                    if (StringUtils.hasText(path)) {
+                        //delete files
+                        filesService.remove(path, UserAuthenticationHelper.getUserAuthentication());
+                    }
+                }
+
+                //delete entity
+                entityService.delete(id);
+            }
         } catch (StoreException e) {
             log.error("store error: {}", e.getMessage());
             throw new SystemException(e.getMessage());
@@ -417,27 +438,41 @@ public class ArtifactServiceImpl
     }
 
     @Override
-    public void deleteArtifacts(@NotNull String project, @NotNull String name) {
+    public void deleteArtifacts(@NotNull String project, @NotNull String name, @Nullable Boolean cascade) {
         log.debug("delete artifacts for project {} with name {}", project, name);
 
-        Specification<ArtifactEntity> spec = Specification.allOf(
-            CommonSpecification.projectEquals(project),
-            CommonSpecification.nameEquals(name)
-        );
-        try {
-            long count = entityService.deleteAll(spec);
-            log.debug("deleted count {}", count);
-        } catch (StoreException e) {
-            log.error("store error: {}", e.getMessage());
-            throw new SystemException(e.getMessage());
+        if (Boolean.TRUE.equals(cascade)) {
+            //delete one by one with cascade
+            findArtifacts(project, name).forEach(a -> deleteArtifact(a.getId(), Boolean.TRUE));
+        } else {
+            //bulk delete entities only
+            Specification<ArtifactEntity> spec = Specification.allOf(
+                CommonSpecification.projectEquals(project),
+                CommonSpecification.nameEquals(name)
+            );
+            try {
+                long count = entityService.deleteAll(spec);
+                log.debug("bulk deleted count {}", count);
+            } catch (StoreException e) {
+                log.error("store error: {}", e.getMessage());
+                throw new SystemException(e.getMessage());
+            }
         }
     }
 
     @Override
-    public void deleteArtifactsByProject(@NotNull String project) {
+    public void deleteArtifactsByProject(@NotNull String project, @Nullable Boolean cascade) {
         log.debug("delete artifacts for project {}", project);
         try {
-            entityService.deleteAll(CommonSpecification.projectEquals(project));
+            if (Boolean.TRUE.equals(cascade)) {
+                //delete one by one with cascade
+                entityService
+                    .searchAll(CommonSpecification.projectEquals(project))
+                    .forEach(a -> deleteArtifact(a.getId(), Boolean.TRUE));
+            } else {
+                //bulk delete entities only
+                entityService.deleteAll(CommonSpecification.projectEquals(project));
+            }
         } catch (StoreException e) {
             log.error("store error: {}", e.getMessage());
             throw new SystemException(e.getMessage());
@@ -503,7 +538,7 @@ public class ArtifactServiceImpl
                 throw new NoSuchEntityException("file");
             }
 
-            DownloadInfo info = filesService.getDownloadAsUrl(path);
+            DownloadInfo info = filesService.getDownloadAsUrl(path, UserAuthenticationHelper.getUserAuthentication());
             if (log.isTraceEnabled()) {
                 log.trace("download url for artifact with id {}: {} -> {}", id, path, info);
             }
@@ -542,7 +577,10 @@ public class ArtifactServiceImpl
                 })
                 .orElse(path);
 
-            DownloadInfo info = filesService.getDownloadAsUrl(fullPath);
+            DownloadInfo info = filesService.getDownloadAsUrl(
+                fullPath,
+                UserAuthenticationHelper.getUserAuthentication()
+            );
             if (log.isTraceEnabled()) {
                 log.trace("download url for artifact with id {} and path {}: {} -> {}", id, sub, path, info);
             }
@@ -583,7 +621,7 @@ public class ArtifactServiceImpl
                     throw new NoSuchEntityException("file");
                 }
 
-                files = filesService.getFileInfo(path);
+                files = filesService.getFileInfo(path, UserAuthenticationHelper.getUserAuthentication());
             }
 
             if (files == null) {
@@ -647,7 +685,7 @@ public class ArtifactServiceImpl
                 }
             }
 
-            UploadInfo info = filesService.getUploadAsUrl(path);
+            UploadInfo info = filesService.getUploadAsUrl(path, UserAuthenticationHelper.getUserAuthentication());
             if (log.isTraceEnabled()) {
                 log.trace("upload url for artifact with id {}: {}", id, info);
             }
@@ -688,7 +726,7 @@ public class ArtifactServiceImpl
                 }
             }
 
-            UploadInfo info = filesService.startMultiPartUpload(path);
+            UploadInfo info = filesService.startMultiPartUpload(path, UserAuthenticationHelper.getUserAuthentication());
             if (log.isTraceEnabled()) {
                 log.trace("start upload url for artifact with id {}: {}", id, info);
             }
@@ -732,7 +770,12 @@ public class ArtifactServiceImpl
                 }
             }
 
-            UploadInfo info = filesService.uploadMultiPart(path, uploadId, partNumber);
+            UploadInfo info = filesService.uploadMultiPart(
+                path,
+                uploadId,
+                partNumber,
+                UserAuthenticationHelper.getUserAuthentication()
+            );
             if (log.isTraceEnabled()) {
                 log.trace("part upload url for artifact with path {}: {}", path, info);
             }
@@ -776,7 +819,12 @@ public class ArtifactServiceImpl
                 }
             }
 
-            UploadInfo info = filesService.completeMultiPartUpload(path, uploadId, eTagPartList);
+            UploadInfo info = filesService.completeMultiPartUpload(
+                path,
+                uploadId,
+                eTagPartList,
+                UserAuthenticationHelper.getUserAuthentication()
+            );
             if (log.isTraceEnabled()) {
                 log.trace("complete upload url for artifact with path {}: {}", path, info);
             }
